% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay_SIR_model_on_data.R
\name{overlay_SIR_model_on_data}
\alias{overlay_SIR_model_on_data}
\title{A function to calculate the deterministic estimates of the number
infected versus time and the average number of descendant infections (ANDI)
for a deterministic Susceptible, Infected, Recovered (SIR) model
and overlay the results on Midwest 2008 influenza outbreak data}
\usage{
overlay_SIR_model_on_data(R0)
}
\arguments{
\item{R0}{the reproduction number}
}
\description{
A function to calculate the deterministic estimates of the number
infected versus time and the average number of descendant infections (ANDI)
for a deterministic Susceptible, Infected, Recovered (SIR) model
and overlay the results on Midwest 2008 influenza outbreak data
}
