% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR_solve_model.R
\name{SIR_solve_model}
\alias{SIR_solve_model}
\title{A function to calculate the deterministic estimates of the number
infected versus time and the average number of descendant infections (ANDI)
for a deterministic Susceptible, Infected, Recovered (SIR) model}
\usage{
SIR_solve_model(N, R0, gamma, fsusc, time_end, delta_t = 0.01)
}
\arguments{
\item{N}{the population size}

\item{R0}{the reproduction number}

\item{gamma}{the recovery rate, in units one over the units of time}

\item{fsusc}{the initial fraction susceptible}

\item{time_end}{the end time of the simulation}

\item{delta_t}{time step to solve model}
}
\value{
list model compartment values vs time and ANDI
}
\description{
A function to calculate the deterministic estimates of the number
infected versus time and the average number of descendant infections (ANDI)
for a deterministic Susceptible, Infected, Recovered (SIR) model
}
