% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIRfunc.R
\name{SIRfunc}
\alias{SIRfunc}
\title{These are the differential equations for a Susceptibe, Infected, Recovered (SIR)
determinstic model, input to the lsoda() method in the deSolve library}
\usage{
SIRfunc(t, x, vparameters)
}
\arguments{
\item{t}{The time}

\item{x}{The current value of the model compartments}

\item{vparameters}{List of the parameters of the model}
}
\value{
The current value of the derivatives of the model
}
\description{
These are the differential equations for a Susceptibe, Infected, Recovered (SIR)
determinstic model, input to the lsoda() method in the deSolve library
}
